fix: harden protocol boundary semantics
Enforce deadline, correlation, framing, and transport invariants across blocking and async owners. Correct BRC profile and codec behavior, seal retry and macro contracts, refresh public API snapshots, and add regression coverage for the reviewed PR 559 failures.

+++ src/error.rs
@@ -189,9 +189,9 @@ pub enum Error {
     #[error("Camera is still moving, position: pan={pan}, tilt={tilt}")]
     CameraMoving {
         /// Current pan position.
-        pan: i16,
+        pan: i32,
         /// Current tilt position.
-        tilt: i16,
+        tilt: i32,
     },
 
     /// Camera has not been properly initialized or powered on.
